pyproject.toml
[tool.poetry]
name = "ravenpython"
version = "0.1.0"
description = ""
authors = ["Andy Green <aj.green07@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
gunicorn = "^20.0.4"
whitenoise = "^5.1.0"
python-decouple = "^3.3"
dj-database-url = "^0.5.0"
psycopg2 = "^2.8.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
